Moved create_schedules to its own file
